+++ docs/CHANGELOG.md
@@ -15,6 +15,9 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Improves
+- Information rules are now read using InformationRulesInput data class, replicate the form of DMS rules.
 
 ## [0.77.3] - 14-05-24
 ### Fixed